% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acf.farima0d0.R
\name{acf.farima0d0}
\alias{acf.farima0d0}
\title{Autocovariance function (ACVF) of FARIMA(0,d,0)}
\usage{
acf.farima0d0(d, h)
}
\arguments{
\item{d}{fractional differencing parameter}

\item{h}{lag to calculate ACVF at}
}
\value{
lag h ACVF of FARIMA(0,d,0)
}
\description{
Autocovariance function (ACVF) of FARIMA(0,d,0)
}
\examples{
acf.farima0d0(1/4, 2)
}
