% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgpois.R
\name{dgpois}
\alias{dgpois}
\title{pdf of generalized poisson}
\usage{
dgpois(x, lambda, eta)
}
\arguments{
\item{x}{series}

\item{lambda}{1 parameter}

\item{eta}{2 parameter}
}
\value{
pdf
}
\description{
pdf of generalized poisson
}
