% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixed_poission_distribution_functions.R
\name{dmixpois}
\alias{dmixpois}
\title{density of mixed-Poisson distribution}
\usage{
dmixpois(x, p, lam1, lam2)
}
\arguments{
\item{x}{evaluation point (or vector)}

\item{p}{mixing probability}

\item{lam1}{mean of first component}

\item{lam2}{mean of second component}
}
\value{
density at x with specified parameters
}
\description{
density of mixed-Poisson distribution
}
