% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farima_sim.R
\name{farima.sim}
\alias{farima.sim}
\title{Simulate from FARIMA(0,d,0)}
\usage{
farima.sim(Tt, d)
}
\arguments{
\item{Tt}{length of the time series to be generated}

\item{d}{long memory parameter}
}
\value{
time series of length Tt
}
\description{
Simulate from FARIMA(0,d,0)
}
\examples{
farima.sim(200, 1/4)
}
