% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hermite_poly.R
\name{Hermite_poly}
\alias{Hermite_poly}
\title{Hermite Polynomials}
\usage{
Hermite_poly(k, x)
}
\arguments{
\item{k}{Hermite polynomial}

\item{x}{input}
}
\value{
value of H_k(x)
}
\description{
evaluate the k-th Hermite Polynomial at x
}
