% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LGC.R
\name{LGC}
\alias{LGC}
\title{Latent Gaussian Count model builder}
\usage{
LGC(x, count.family = c("Poisson", "mixed-Poisson", "negbinom"),
  gauss.series = c("AR", "FARIMA"), estim.method = c("gaussianLik",
  "particlesSIS"), max.terms = 30, p = NULL, d = NULL, q = NULL,
  n.mix = NULL, n = NULL, print.progress = FALSE,
  print.initial.estimates = FALSE, ...)
}
\arguments{
\item{x}{data}

\item{count.family}{desired marginal distribution (Poisson,mixed-Poisson,etc)}

\item{gauss.series}{desired structure of your latent Gaussian process}

\item{estim.method}{method used for estimating parameters}

\item{max.terms}{maximum number of terms used to truncate Hermite expansions}

\item{p}{AR order}

\item{d}{don't use it}

\item{q}{don't use it}

\item{n.mix}{number of Poisson distributions in mixed-Poisson count.family}

\item{n}{Not sure what this is}

\item{print.progress}{Should progress be printed as optim() is run}

\item{print.initial.estimates}{Should initial estiamtes be printed}

\item{...}{additional parameters to pass to optim}
}
\value{
list object which are results from optim() run
}
\description{
Main execution function for latentGaussCounts package.
}
