% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixed_poission_distribution_functions.R
\name{mixpois.ar1}
\alias{mixpois.ar1}
\title{Generate simulated mixed-Poisson-AR1 model}
\usage{
mixpois.ar1(Tt, phi, p, lam1, lam2)
}
\arguments{
\item{Tt}{length of the series}

\item{phi}{AR(1) parameter}

\item{p}{mixture probability}

\item{lam1}{mean of first component}

\item{lam2}{mean of second component}
}
\value{
simulated series of length Tt
}
\description{
Generate simulated mixed-Poisson-AR1 model
}
