% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgpois.R
\name{pgpois}
\alias{pgpois}
\title{cdf of generalized poisson}
\usage{
pgpois(x, lambda, eta)
}
\arguments{
\item{x}{series}

\item{lambda}{1 parameter}

\item{eta}{2 parameter}
}
\value{
cdf
}
\description{
cdf of generalized poisson
}
