% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixed_poission_distribution_functions.R
\name{pmixpois}
\alias{pmixpois}
\title{CDF of mixed-Poisson distribution}
\usage{
pmixpois(x, p, lam1, lam2)
}
\arguments{
\item{x}{evaluation point (or vector)}

\item{p}{mixing probability}

\item{lam1}{mean of first component}

\item{lam2}{mean of second component}
}
\value{
CDF at x with specified parameters
}
\description{
CDF of mixed-Poisson distribution
}
