% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixed_poission_distribution_functions.R
\name{pmixpois_vecparam}
\alias{pmixpois_vecparam}
\title{Vector input CDF of mixed-Poisson distribution}
\usage{
pmixpois_vecparam(x, param)
}
\arguments{
\item{x}{evaluation point (or vector)}

\item{param}{vector of parameters. c(p, lam1, lam2)}
}
\value{
CDF at x with specified parameters
}
\description{
Vector input CDF of mixed-Poisson distribution
}
