% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pois_farima.R
\name{pois.farima}
\alias{pois.farima}
\title{Simulate from Poisson-FARIMA(0,d,0) model}
\usage{
pois.farima(Tt, d, lam)
}
\arguments{
\item{Tt}{length of series to be generated}

\item{d}{long-memory parameter}

\item{lam}{poisson parameter}
}
\value{
discrete valued time series with Poisson marginals and FARIMA(0,d,0)
    covariance structure
}
\description{
Simulate from Poisson-FARIMA(0,d,0) model
}
\examples{
pois.farima(200, 1/4, 3)
}
