% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixed_poission_distribution_functions.R
\name{qmixpois}
\alias{qmixpois}
\title{Inverse CDF/Quantile function for mixed-Poisson distribution}
\usage{
qmixpois(y, p, lam1, lam2)
}
\arguments{
\item{y}{evaluation point (or vector)}

\item{p}{mixing probability}

\item{lam1}{mean of first component}

\item{lam2}{mean of second component}
}
\value{
quantile function at y with specified parameters
}
\description{
Inverse CDF/Quantile function for mixed-Poisson distribution
}
