% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixed_poission_distribution_functions.R
\name{rmixpois}
\alias{rmixpois}
\title{Generate random mixed-Poisson draws}
\usage{
rmixpois(n, p, lam1, lam2)
}
\arguments{
\item{n}{number of observations to generate}

\item{p}{mixing probability}

\item{lam1}{mean of first component}

\item{lam2}{mean of second component}
}
\value{
vector of randomly generated values
}
\description{
Generate random mixed-Poisson draws
}
