% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_boxplots.R
\name{sim_boxplots}
\alias{sim_boxplots}
\title{Simulation boxplot function}
\usage{
sim_boxplots(data)
}
\arguments{
\item{data}{A dataframe with columns "parameter" (a character eg "lambda")
"estimation.method" (eg "Gaussian-lik")
 "value" (numeric paramter estimates)}
}
\value{
a boxplot
}
\description{
Simulation boxplot function
}
