% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_mixedpois_ar1.R
\name{sim_mixedpois_ar1}
\alias{sim_mixedpois_ar1}
\title{Simulate from mixed-Poisson-AR1 model}
\usage{
sim_mixedpois_ar1(n, phi, p, lam1, lam2)
}
\arguments{
\item{n}{number of observations}

\item{phi}{AR parameter}

\item{p}{mixing probability}

\item{lam1}{mean of first Poisson component}

\item{lam2}{mean of second Poisson component}
}
\value{
simulated mixed-Poisson-AR1 model of length n
}
\description{
Simulate from mixed-Poisson-AR1 model
}
