% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_negbinom_ma1.R
\name{sim_negbinom_ma1}
\alias{sim_negbinom_ma1}
\title{Simulate from NegativeBinomial-MA model}
\usage{
sim_negbinom_ma1(n, theta, r, p)
}
\arguments{
\item{n}{length of series}

\item{theta}{MA(1) parameter}

\item{r}{number of failures before x successes}

\item{p}{success probability}
}
\value{
simulated count series of length n
}
\description{
Simulate from NegativeBinomial-MA model
}
