% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_pois_ar.R
\name{sim_pois_ar}
\alias{sim_pois_ar}
\title{Simulate from Poisson-AR1 model}
\usage{
sim_pois_ar(n, phi, lam)
}
\arguments{
\item{n}{length of series}

\item{phi}{AR(1) parameter}

\item{lam}{Poisson parameter}
}
\value{
simulated count series of length n
}
\description{
Simulate from Poisson-AR1 model
}
