% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_pois_ma.R
\name{sim_pois_ma}
\alias{sim_pois_ma}
\title{Simulate from Poisson-MA model}
\usage{
sim_pois_ma(n, theta, lam)
}
\arguments{
\item{n}{length of series}

\item{lam}{Poisson parameter}

\item{phi}{MA parameter}
}
\value{
simulated count series of length n
}
\description{
Simulate from Poisson-MA model
}
